tests/wpt/tests/dom/nodes/ParentNode-querySelector-case-insensitive.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>querySelector(All) must work with the i and *= selectors</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/2551 -->

<input name="User" id="testInput"></input>

<script>
"use strict";
const input = document.getElementById("testInput");

test(() => {
  assert_equals(document.querySelector("input[name*=user i]"), input);
}, "querySelector");

test(() => {
  assert_array_equals(document.querySelectorAll("input[name*=user i]"), [input]);
}, "querySelectorAll");
</script>
